+++ crates/render/vertex_pulling/index_buffer.rs
@@ -26,7 +26,7 @@ const NUM_CUBE_INDICES_USIZE: usize = 3 * 3 * 2;
 /// to indicate which of the 3 faces is being rendered.
 #[rustfmt::skip]
 #[allow(clippy::unusual_byte_groupings)]
-pub(crate) const CUBE_INDICES: [u32; NUM_CUBE_INDICES_USIZE] = [
+pub(crate) const CUBE_INDICES: [u16; NUM_CUBE_INDICES_USIZE] = [
     0b00_000, 0b00_010, 0b00_001, 0b00_010, 0b00_011, 0b00_001, // face XY (0)
     0b01_101, 0b01_100, 0b01_001, 0b01_001, 0b01_100, 0b01_000, // face XZ (1)
     0b10_000, 0b10_100, 0b10_110, 0b10_000, 0b10_110, 0b10_010, // face YZ (2)
